docs: Add press section (#12)
This pull request adds a new documentation page to highlight media
mentions of Datavzrd. The page includes a table summarizing selected
mentions with links to articles and institutional websites.

### Documentation updates:

*
[`src/docs/press.rst`](diffhunk://#diff-98185d0426ace1c1b9fa9dfaa7b8316f5024bea94579acb3e26f2b43dbd90e21R1-R14):
Added a new "Datavzrd in the Press" section, including a table with
sources and links to media coverage, such as Universität Duisburg-Essen
and Science News Today. The page also invites users to share additional
mentions of Datavzrd.